backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.7.0
Pillow==11.1.0
python-decouple==3.8
django-filter==24.2
gunicorn==22.0.0
whitenoise==6.7.0 